requirements.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
alembic==1.13.1
httpx==0.27.0
google-generativeai==0.3.2
anthropic==0.8.1
openai==1.12.0 